fix(analysis): classify locals, fields and semantic token kinds correctly
- Capture class fields as fields in LOCALS_QUERY instead of also capturing them as variables, so a field is no longer both at once and the unused-variable rule stops reporting them
- Keep module-scope and namespace-scope declarations out of the unused-variable rule: "unused" is not decidable from a single document for a declaration another file or the engine may reference
- Resolve identifier references through the scope tree before emitting semantic tokens: the highlights query is purely syntactic and cannot tell a parameter read from a field read from a local read, so all three used to come back as a plain variable
- Stop deducing a type from C-style integer suffixes in inlay hints; AngelScript has no such suffixes
- Add SemanticTokenKindTest, plus coverage for field/local classification in the scope collector, document highlights and the semantic analyzer

+++ server/src/features/formatting/FormattingHandler.cpp
@@ -1159,7 +1159,9 @@ namespace angel_lsp::features
 
         lsp::TextEdit edit;
         edit.range.start = lsp::Position{ 0, 0 };
-        edit.range.end = lsp::Position{ lineCount, lastLineLen };
+        // Narrowing: size_t is 64-bit on x64, so the braced initialiser is a hard error there
+        // even though the 32-bit build accepts it silently.
+        edit.range.end = lsp::Position{ lineCount, static_cast<uint32_t>(lastLineLen) };
         edit.newText = std::move(formatted);
 
         return std::vector<lsp::TextEdit>{ std::move(edit) };
